requirements.txt
openai-whisper
ffmpeg-python
opencv-python
pandas
numpy
moviepy
torch
transformers
jupyter
notebook
scikit-learn
matplotlib
